Make unplate private
