+++ pyro/primitives.py
@@ -11,7 +11,6 @@ import pyro.distributions as dist
 import pyro.infer as infer
 import pyro.poutine as poutine
 from pyro.params import param_with_module_name
-from pyro.poutine.messenger import mute_messengers
 from pyro.poutine.plate_messenger import PlateMessenger
 from pyro.poutine.runtime import _MODULE_NAMESPACE_DIVIDER, _PYRO_PARAM_STORE, am_i_wrapped, apply_stack, effectful
 from pyro.poutine.subsample_messenger import SubsampleMessenger
@@ -319,56 +318,6 @@ def plate_stack(prefix, sizes, rightmost_dim=-1):
         for i, size in enumerate(reversed(sizes)):
             plate_i = plate("{}_{}".format(prefix, i), size, dim=rightmost_dim - i)
             stack.enter_context(plate_i)
-        yield
-
-
-@contextmanager
-def unplate(name=None, dim=None):
-    """
-    Context manager to temporarily exit a single enclosing plate.
-
-    This is useful for sampling auxiliary variables or lazily sampling global
-    variables that are needed in a plated context. For example the following
-    models are equivalent:
-
-    Example::
-
-        def model_1(data):
-            loc = pyro.sample("loc", dist.Normal(0, 1))
-            with pyro.plate("data", len(data)):
-                with pyro.unplate("data"):
-                    scale = pyro.sample("scale", dist.LogNormal(0, 1))
-                pyro.sample("x", dist.Normal(loc, scale))
-
-        def model_2(data):
-            loc = pyro.sample("loc", dist.Normal(0, 1))
-            scale = pyro.sample("scale", dist.LogNormal(0, 1))
-            with pyro.plate("data", len(data)):
-                pyro.sample("x", dist.Normal(loc, scale))
-
-    :param str name: Optional name of plate to match.
-    :param int dim: Optional dim of plate to match. Must be negative.
-    :raises: ValueError if no enclosing plate was found.
-    """
-    if (name is not None) == (dim is not None):
-        raise ValueError("Exactly one of name,dim must be specified")
-    if name is not None:
-        assert isinstance(name, str)
-    if dim is not None:
-        assert isinstance(dim, int)
-        assert dim < 0
-
-    def predicate(messenger):
-        if not isinstance(messenger, PlateMessenger):
-            return False
-        if name is not None:
-            return messenger.name == name
-        if dim is not None:
-            return messenger.dim == dim
-
-    with mute_messengers(predicate) as matches:
-        if len(matches) != 1:
-            raise ValueError("unplate matched {} messengers".format(len(matches)))
         yield
