+++ tests/infer/test_valid_models.py
@@ -20,6 +20,7 @@ from pyro.infer.tracetmc_elbo import TraceTMC_ELBO
 from pyro.infer.util import torch_item
 from pyro.ops.indexing import Vindex
 from pyro.optim import Adam
+from pyro.poutine.plate_messenger import unplate
 from tests.common import assert_close
 
 logger = logging.getLogger(__name__)
@@ -867,7 +868,7 @@ def test_unplate_name_ok():
         with pyro.plate("plate", 2):
             b = pyro.sample("b", dist.Normal(0, 1))
             assert b.shape == (2,)
-            with pyro.unplate("plate"):
+            with unplate("plate"):
                 c = pyro.sample("c", dist.Normal(0, 1))
                 assert c.shape == ()
 
@@ -877,7 +878,7 @@ def test_unplate_name_ok():
         with pyro.plate("plate", 2):
             b = pyro.sample("b", dist.Normal(0, 1))
             assert b.shape == (2,)
-            with pyro.unplate("plate"):
+            with unplate("plate"):
                 a = pyro.sample("a", dist.Normal(0, 1))
                 assert a.shape == ()
 
@@ -892,7 +893,7 @@ def test_unplate_dim_ok():
         with pyro.plate("plate", 2):
             b = pyro.sample("b", dist.Normal(0, 1))
             assert b.shape == (2,)
-            with pyro.unplate(dim=-1):
+            with unplate(dim=-1):
                 c = pyro.sample("c", dist.Normal(0, 1))
                 assert c.shape == ()
 
@@ -902,7 +903,7 @@ def test_unplate_dim_ok():
         with pyro.plate("plate", 2):
             b = pyro.sample("b", dist.Normal(0, 1))
             assert b.shape == (2,)
-            with pyro.unplate(dim=-1):
+            with unplate(dim=-1):
                 a = pyro.sample("a", dist.Normal(0, 1))
                 assert a.shape == ()
 
@@ -912,7 +913,7 @@ def test_unplate_dim_ok():
 def test_unplate_missing_error():
 
     def model():
-        with pyro.unplate("plate"):
+        with unplate("plate"):
             pyro.sample("a", dist.Normal(0, 1))
 
     def guide():
